cmd/migration-managerd/internal/api: Move source sync helpers to their own file
